+++ bundles/org.openhab.binding.logreader/src/3rdparty/java/org/apache/commons/io/input/TailerListener.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.openhab.binding.logreader.internal.commonsio;
+package org.apache.commons.io.input;
 
 /**
  * Listener for events from a {@link Tailer}.
